Create `cached_tree_hash` crate.
