+++ eth2/utils/cached_tree_hash/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "cached_tree_hash"
+version = "0.1.0"
+authors = ["Paul Hauner <paul@paulhauner.com>"]
+edition = "2018"
+
+[dev-dependencies]
+tree_hash_derive = { path = "../tree_hash_derive" }
+
+[dependencies]
+tree_hash = { path = "../tree_hash" }
+ethereum-types = "0.5"
+hashing = { path = "../hashing" }
+int_to_bytes = { path = "../int_to_bytes" }
