+++ eth2/utils/cached_tree_hash/src/errors.rs
@@ -1,0 +1,18 @@
+use tree_hash::TreeHashType;
+
+#[derive(Debug, PartialEq, Clone)]
+pub enum Error {
+    ShouldNotProduceBTreeOverlay,
+    NoFirstNode,
+    NoBytesForRoot,
+    UnableToObtainSlices,
+    UnableToGrowMerkleTree,
+    UnableToShrinkMerkleTree,
+    TreeCannotHaveZeroNodes,
+    ShouldNeverBePacked(TreeHashType),
+    BytesAreNotEvenChunks(usize),
+    NoModifiedFieldForChunk(usize),
+    NoBytesForChunk(usize),
+    NoOverlayForIndex(usize),
+    NotLeafNode(usize),
+}
