+++ eth2/utils/cached_tree_hash/src/resize.rs
